--- src/store/index.ts
@@ -1,2 +1,0 @@
-export { useCartStore } from "./cart-store";
-export type { CartItem } from "./cart-store";
