+++ src/types/cart.ts
@@ -1,0 +1,7 @@
+export interface CartItem {
+  id: number;
+  title: string;
+  price: number;
+  image: string;
+  quantity: number;
+}
